+++ installer/compile.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import sys 
 import json
+import time
 
 from jinja2 import Environment, PackageLoader, select_autoescape
 import requests
@@ -85,6 +86,7 @@ def get_vars(input_data, default_data):
                      'deploy_ckn_mqtt': True,
                      'deploy_power_monitoring': False,
                      'deploy_oracle': False,
+                     'motion_video_device': '/dev/video0',
                      'inference_server': True}
     simulation_defaults = {'deploy_image_generating': True,
                            'deploy_image_detecting': False,
@@ -92,12 +94,24 @@ def get_vars(input_data, default_data):
                            'deploy_ckn': True,
                            'deploy_ckn_mqtt': False,
                            'deploy_oracle': True,
+                           'expanded_metrics': True,
                            'inference_server': False}
+    video_simulation_defaults = {'deploy_image_generating': False,
+                                 'deploy_video_generating': True,
+                                 'deploy_image_detecting': True,
+                                 'deploy_reporter': True,
+                                 'deploy_ckn': False,
+                                 'deploy_ckn_mqtt': False,
+                                 'deploy_oracle': False,
+                                 'use_bundled_example_images': False,
+                                 'inference_server': False}
 
     if vars.get("mode") == 'demo':
         vars = { **default_data, **demo_defaults, **input_data }
     elif vars.get("mode") == 'simulation':
         vars = { **default_data, **simulation_defaults, **input_data }
+    elif vars.get("mode") == 'video_simulation':
+        vars = { **default_data, **video_simulation_defaults, **input_data }
 
     # the powerjoular backend requires the docker socket to function:
     if vars.get("power_monitor_backend") == 'powerjoular':
@@ -143,6 +157,18 @@ def get_vars(input_data, default_data):
         if vars.get("model_id") is None:
             vars['model_id'] = '41d3ed40-b836-4a62-b3fb-67cee79f33d9-model'
 
+    if vars.get('mode') == 'video_simulation':
+        # for video simulations, determine if motion is using device or netcam
+        if vars.get('motion_video_device'):
+            vars['motion_video_type'] = 'device'
+        else:
+            vars['motion_video_type'] = 'file'
+        # determine which of url, local file, and example file are used for the video
+        if vars.get('source_video_url') or vars.get('local_video_path'):
+            vars['use_example_video'] = False
+        elif vars.get('use_example_video'):
+            vars['local_video_path'] = './video.mp4'
+
     # Add the installer's UID and GID
     vars["uid"] = uid
     vars["gid"] = gid 
@@ -164,16 +190,22 @@ def get_vars(input_data, default_data):
     print(f"Merged variables: {vars}")
     return vars 
 
-def get_urls_from_ckn(model_id):
+def get_urls_from_ckn(endpoint, model_id):
     """
     Given a model card ID, extract the download URL and inference labels URL.
     """
     if model_id.endswith("-model"):
         model_id = model_id[:-6]
-    patra_download_endpoint = f"https://ckn.d2i.tacc.cloud/patra/download_mc?id={model_id}"
+    patra_download_endpoint = f"{endpoint}?id={model_id}"
 
-    response = requests.get(patra_download_endpoint)
-    if response.status_code != 200:
+    num_tries = 0
+    while num_tries < 5:
+        num_tries = num_tries + 1
+        response = requests.get(patra_download_endpoint)
+        if response.status_code == 200:
+            break
+        time.sleep(2)
+    else:
         raise Exception(f"Failed to fetch data. Status code: {response.status_code}")
 
     try:
@@ -204,7 +236,7 @@ def download_model_by_id(vars, full_install_dir):
     model_url = None
     label_url = None
     print(f"Checking CKN for the URL to the pt file...")
-    model_url, label_url = get_urls_from_ckn(model_id)
+    model_url, label_url = get_urls_from_ckn(vars['patra_endpoint'], model_id)
     print(f"Got URL for model from CKN; URL: {model_url}")
     print(f"Got URL for labels from CKN; URL: {label_url}")
     # if we have a model URL, then we download it so it can be mounted 
@@ -309,6 +341,30 @@ def generate_additional_directories(vars, full_install_dir):
             sys.exit(1)
         print(f"Using URL for images: {vars['source_image_url']}")
 
+    if vars['use_example_video'] == True: 
+        try:
+            shutil.copy("/defaults/example_video/example_video.mp4", os.path.join(full_install_dir, "video.mp4"))
+            shutil.copy("/defaults/example_video/ground_truth.yml", os.path.join(full_install_dir, "ground_truth.yml"))
+        except Exception as e:
+            print(f"ERROR: Could not copy bundled example video; error: {e}")
+            print("Exiting...")
+            sys.exit(1)
+    else:
+        if vars['source_video_url']:
+            try:
+                rsp = requests.get(vars['source_video_url'])
+                rsp.raise_for_status()
+            except Exception as e:
+                print(f'Error: could not download video at URL {source_video_url}; details: {e}')
+                sys.exit(1)
+            video_install_path = os.path.join(full_install_dir, 'video.mp4')
+            with open(video_install_path, 'wb') as f:
+                f.write(rsp.content)
+        elif vars['local_video_path']:
+            full_video_path = os.path.join(full_install_dir, vars.get('local_video_path'))
+            if not os.path.exists(full_video_path):
+                print(f'ERROR: local_video_path must be a relative path to the install directory and must already exist; the computed path ({full_video_dir}) does not exist.\nExiting...')
+                sys.exit(1)
 
     # create output directories if they do not exist     
     images_output_dir = os.path.join(full_install_dir, vars["images_output_dir"])
@@ -327,6 +383,11 @@ def generate_additional_directories(vars, full_install_dir):
     detection_output_dir = os.path.join(full_install_dir, vars["detection_reporter_plugin_output_dir"])
     if not os.path.exists(detection_output_dir):
         os.makedirs(detection_output_dir)
+
+    if vars['deploy_video_generating']:
+        video_output_dir = os.path.join(full_install_dir, vars["video_output_dir"])
+        if not os.path.exists(video_output_dir):
+            os.makedirs(video_output_dir)
     
 
 def main():